CMake: Resolve #187 by preferring external GLEW:
On systems that have GLEW installed system-wide (or in another way
that makes it end up in an include directory we use), compiling the
internal GLEW may mix in the system GLEW, which may be an incompatible
version.

With this change, by default, system GLEW (found via find_package(GLEW))
will be preferred, and the OF internal glew will only be used if either
OPENFLIPPER_FORCE_INTERNAL_GLEW is set, or find_package fails and
OPENFLIPPER_DISABLE_INTERNAL_GLEW is not set.